+++ lib/Dialog/addTodoDialog.dart
@@ -1,0 +1,50 @@
+import 'package:flutter/material.dart';
+
+class AddTodoDialog extends StatefulWidget {
+  final Function(String) onAddTodo;
+
+  const AddTodoDialog({super.key, required this.onAddTodo});
+
+  @override
+  State<AddTodoDialog> createState() => _AddTodoDialogState();
+}
+
+class _AddTodoDialogState extends State<AddTodoDialog> {
+  final TextEditingController _controller = TextEditingController();
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: const Text("Add Todo"),
+      content: TextField(
+        controller: _controller,
+        decoration: const InputDecoration(hintText: "タイトル入力"),
+      ),
+      actions: [
+        TextButton(
+          onPressed: () {
+            Navigator.of(context).pop();
+          },
+          child: const Text("キャンセル"),
+        ),
+        ElevatedButton(
+          onPressed: () {
+            print("Todoが追加されました");
+            print("Todoの内容: ${_controller.text}");
+            if (_controller.text.isNotEmpty) {
+              widget.onAddTodo(_controller.text);
+            }
+            Navigator.of(context).pop();
+          },
+          child: const Text("追加"),
+        ),
+      ],
+    );
+  }
+}
