+++ lib/main.dart
@@ -1,3 +1,5 @@
+import 'package:first_app/Dialog/addTodoDialog.dart';
+import 'package:first_app/Dialog/editTodoDialog.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -142,115 +144,6 @@ class TodoList extends StatelessWidget {
               ],
             ),
           ),
-      ],
-    );
-  }
-}
-
-class AddTodoDialog extends StatefulWidget {
-  final Function(String) onAddTodo;
-
-  const AddTodoDialog({super.key, required this.onAddTodo});
-
-  @override
-  State<AddTodoDialog> createState() => _AddTodoDialogState();
-}
-
-class _AddTodoDialogState extends State<AddTodoDialog> {
-  final TextEditingController _controller = TextEditingController();
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      title: const Text("Add Todo"),
-      content: TextField(
-        controller: _controller,
-        decoration: const InputDecoration(hintText: "タイトル入力"),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () {
-            Navigator.of(context).pop();
-          },
-          child: const Text("キャンセル"),
-        ),
-        ElevatedButton(
-          onPressed: () {
-            print("Todoが追加されました");
-            print("Todoの内容: ${_controller.text}");
-            if (_controller.text.isNotEmpty) {
-              widget.onAddTodo(_controller.text);
-            }
-            Navigator.of(context).pop();
-          },
-          child: const Text("追加"),
-        ),
-      ],
-    );
-  }
-}
-
-class EditTodoDialog extends StatefulWidget {
-  final int index;
-  final String currentTodo;
-  final Function(int, String) onEditTodo;
-  const EditTodoDialog({
-    super.key,
-    required this.index,
-    required this.currentTodo,
-    required this.onEditTodo,
-  });
-  @override
-  State<EditTodoDialog> createState() => _EditTodoDialogState();
-}
-
-class _EditTodoDialogState extends State<EditTodoDialog> {
-  final TextEditingController _controller = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    _controller.text = widget.currentTodo;
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      title: const Text("Edit Todo"),
-      content: TextField(
-        controller: _controller,
-        decoration: const InputDecoration(hintText: "タイトル入力"),
-      ),
-      actions: [
-        TextButton(
-          onPressed: () {
-            Navigator.of(context).pop();
-          },
-          child: const Text("キャンセル"),
-        ),
-        ElevatedButton(
-          onPressed: () {
-            print("Todoが変更されました");
-            print("Todoの内容: ${_controller.text}");
-            if (_controller.text.isNotEmpty) {
-              widget.onEditTodo(widget.index, _controller.text);
-            }
-            Navigator.of(context).pop();
-          },
-          child: const Text("変更"),
-        ),
       ],
     );
   }
